:recycle: aggregate config declarations
